Move typed writer/reader to blob package
