+++ game/asset/binary_v1.go
@@ -3,11 +3,11 @@ package asset
 import (
 	"io"
 
-	"github.com/mokiat/lacking/data/storage"
+	"github.com/mokiat/lacking/util/blob"
 )
 
 func (b *Binary) encodeV1(out io.Writer) error {
-	writer := storage.NewTypedWriter(out)
+	writer := blob.NewTypedWriter(out)
 	if err := writer.WriteByteBlock(b.Data); err != nil {
 		return err
 	}
@@ -15,7 +15,7 @@ func (b *Binary) encodeV1(out io.Writer) error {
 }
 
 func (b *Binary) decodeV1(in io.Reader) error {
-	reader := storage.NewTypedReader(in)
+	reader := blob.NewTypedReader(in)
 	if data, err := reader.ReadBytesBlock(); err != nil {
 		return err
 	} else {
